SLIME Gravity
I was rejected from participating in the SAT Competition 2022, because I am self-taught.
This is the SLIME-Gravity code, the version with which I would participate in this competition.
Here the report where it is clearly noted that it is better than Kissat MAB on a large scale.
If you have a real problem, SLIME is your best option.